pyproject.toml
[tool.poetry]
name = "bloatbuster"
version = "0.1.0"
description = "A tool to identify and manage unnecessary files for Dropbox sync"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
loguru = "^0.5.3"
tqdm = "^4.62.3"
pyyaml = "^6.0"
maturin = "^0.13.4"  # Specific version known to work

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
black = "^21.5b1"
isort = "^5.8.0"
mypy = "^0.920"
bandit = "^1.7.0"
safety = "^1.10.3"
pytest = "^6.2.4"
tox = "^3.28.0"
sphinx = "^3.5.4"
mkdocs = "^1.2.2"
pre-commit = "^2.13.0"
coverage = "^5.5"
setuptools = "^65.5.0"
wheel = "^0.37.1"
typed-ast = "^1.5.5"
pylint = "^2.13.9"
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
